+++ Source/EventFlow.Sql/ReadModels/SqlReadModelStore.cs
@@ -52,7 +52,6 @@ namespace EventFlow.Sql.ReadModels
         private readonly ITransientFaultHandler<IOptimisticConcurrencyRetryStrategy> _transientFaultHandler;
         private static readonly Func<TReadModel, int?> GetVersion;
         private static readonly Action<TReadModel, int?> SetVersion;
-        private static readonly Func<TReadModel, string> GetIdentity;
         private static readonly Action<TReadModel, string> SetIdentity;
 
         static SqlReadModelStore()
@@ -77,12 +76,10 @@ namespace EventFlow.Sql.ReadModels
                 .SingleOrDefault(p => p.GetCustomAttributes().Any(a => a is SqlReadModelIdentityColumnAttribute));
             if (identityPropertyInfo == null)
             {
-                GetIdentity = rm => null as string;
                 SetIdentity = (rm, i) => { };
             }
             else
             {
-                GetIdentity = rm => (string)identityPropertyInfo.GetValue(rm);
                 SetIdentity = (rm, i) => identityPropertyInfo.SetValue(rm, i);
             }
         }
@@ -113,7 +110,7 @@ namespace EventFlow.Sql.ReadModels
 
                 await _transientFaultHandler.TryAsync(
                     c => UpdateReadModelAsync(readModelContext, updateReadModel, c, readModelUpdate),
-                    Label.Named($"sqlite-read-model-update"),
+                    Label.Named("sql-read-model-update"),
                     cancellationToken)
                     .ConfigureAwait(false);
             }
@@ -133,7 +130,7 @@ namespace EventFlow.Sql.ReadModels
             if (readModel == null)
             {
                 readModel = await _readModelFactory.CreateAsync(readModelUpdate.ReadModelId, cancellationToken).ConfigureAwait(false);
-                readModelEnvelope = ReadModelEnvelope<TReadModel>.With(readModelUpdate.ReadModelId, readModel);
+                readModelEnvelope = ReadModelEnvelope<TReadModel>.With(readModelUpdate.ReadModelId, readModel, false);
             }
 
             var originalVersion = readModelEnvelope.Version;
@@ -143,10 +140,11 @@ namespace EventFlow.Sql.ReadModels
                 readModelEnvelope,
                 cancellationToken)
                 .ConfigureAwait(false);
+            if (!readModelEnvelope.IsModified) return;
 
             if (readModelContext.IsMarkedForDeletion)
             {
-                await DeleteAsync(readModelUpdate.ReadModelId, cancellationToken);
+                await DeleteAsync(readModelUpdate.ReadModelId, cancellationToken).ConfigureAwait(false);
                 return;
             }
 
@@ -203,8 +201,8 @@ namespace EventFlow.Sql.ReadModels
             Log.Verbose(() => $"Found SQL read model '{readModelType.PrettyPrint()}' with ID '{readModelVersion}'");
 
             return readModelVersion.HasValue
-                ? ReadModelEnvelope<TReadModel>.With(id, readModel, readModelVersion.Value)
-                : ReadModelEnvelope<TReadModel>.With(id, readModel);
+                ? ReadModelEnvelope<TReadModel>.With(id, readModel, readModelVersion.Value, false)
+                : ReadModelEnvelope<TReadModel>.With(id, readModel, false);
         }
 
         public override async Task DeleteAsync(
@@ -215,7 +213,7 @@ namespace EventFlow.Sql.ReadModels
             var readModelName = typeof(TReadModel).Name;
 
             var rowsAffected = await _connection.ExecuteAsync(
-                Label.Named("mssql-delete-read-model", readModelName),
+                Label.Named("sql-delete-read-model", readModelName),
                 cancellationToken,
                 sql,
                 new { EventFlowReadModelId = id })